apply-addons: move to new package: cmd/wksctl/applyaddons
