+++ cmd/wksctl/main.go
@@ -9,6 +9,7 @@ import (
 	"github.com/weaveworks/go-checkpoint"
 	"github.com/weaveworks/wksctl/cmd/wksctl/addon"
 	"github.com/weaveworks/wksctl/cmd/wksctl/apply"
+	"github.com/weaveworks/wksctl/cmd/wksctl/applyaddons"
 	"github.com/weaveworks/wksctl/cmd/wksctl/bashcompletions"
 	"github.com/weaveworks/wksctl/cmd/wksctl/kubeconfig"
 	"github.com/weaveworks/wksctl/cmd/wksctl/plan"
@@ -43,6 +44,7 @@ func main() {
 
 	rootCmd.AddCommand(addon.Cmd)
 	rootCmd.AddCommand(apply.Cmd)
+	rootCmd.AddCommand(applyaddons.Cmd)
 	rootCmd.AddCommand(kubeconfig.Cmd)
 	rootCmd.AddCommand(plan.Cmd)
 	rootCmd.AddCommand(registrysynccommands.Cmd)
